pyproject.toml
[project]
name = "assignment-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "warp-lang>=1.10.1",
]
